+++ tests/read.rs
@@ -91,6 +91,7 @@ fn read_all_atomic_types_ok() {
 mod struct_test_1 {
     use super::parser::{Parser, Reader};
     use super::ply;
+    use super::ply::PropertyAccessResult;
     use super::{GREG_TURK_EXAMPLE1_OK_ASCII, read_from_bytes};
     #[derive(Debug)]
     struct Vertex {
@@ -112,13 +113,14 @@ mod struct_test_1 {
                 z: 0.0,
             }
         }
-        fn set_property(&mut self, key: &str, property: ply::Property) {
+        fn set_property(&mut self, key: &str, property: ply::Property) -> PropertyAccessResult {
             match (key, property) {
                 ("x", ply::Property::Float(v)) => self.x = v,
                 ("y", ply::Property::Float(v)) => self.y = v,
                 ("z", ply::Property::Float(v)) => self.z = v,
                 (k, _) => panic!("Vertex: Unexpected key/value combination: key: {}", k),
             }
+            PropertyAccessResult::Set
         }
     }
 
@@ -129,11 +131,12 @@ mod struct_test_1 {
                 vertex_index: Vec::new(),
             }
         }
-        fn set_property(&mut self, key: &str, property: ply::Property) {
+        fn set_property(&mut self, key: &str, property: ply::Property) -> PropertyAccessResult {
             match (key, property) {
                 ("vertex_index", ply::Property::ListInt(vec)) => self.vertex_index = vec,
                 (k, _) => panic!("Face: Unexpected key/value combination: key: {}", k),
             }
+            PropertyAccessResult::Set
         }
     }